Add support for mirror BRDF
* Add a variation on the Cornell box to test mirrors.
* Add support for ensuring roughness never reduces when extending eye
paths.  (For now we can special case this to turn mirror surfaces into
diffuse if we have already seen a diffuse surface in this path.)  This
reduces noise in indirect lighting, at the cost of changing indirect
lighting to be less sharp (in particular, caustic paths are no longer
possible).
* Add UI for max number of bounces and whether to enable keeping
roughness at the maximum seen so far along eye paths.