Remove callable BRDFs, large net perf loss
Pass goes from 30ms to 50ms just by loading the BRDF callable shaders
(they don't have to be called, they just need to be compiled into the
pipeline). NSight shows a large increase in VRAM throughput for the
slow case, so I would guess some registers are getting spilled to
memory.

Could revisit on hardware with a larger register file (tested on a 2060
here), but will just remove the code for now and stick with a dumb
switch statement for BSDFs.