Convert spectral reflectance with illuminant
My understanding is that RGB reflectance values are the response of the
surface under a D65 illuminant, so use this when converting to avoid
having to do E-to-D65 white point conversion afterwards.

It's unclear what to do with light spectral power distributions. Should
these be normalized in some way against D65 for rendering using Rec709
primaries?  Have left as-is for now.

Also boosted the exposure of the cornell box image.