Various sync improvements
* Categorise resource use as either read or write (could extend with
atomic in future). If a resource is used in more than one category,
ensure that there are barriers between work, even if resource usage
does not change between them, to avoid them overlapping.
* Add HOST_READ buffer usage, generalise how final usage works for
swapchains to allow any imported resource to end up in a fixed state,
use this to add appropriate barriers before the capture buffer is read
back by the CPU.
* Add some missing memory flush/invalidate for the capture buffer.
* Replace device wait for idle with more minimal wait for fence.