Refactor async loading to use futures
Replace the system for static resources with futures that can make use
of async/await syntax. Also add a system for futures to be able to
schedule and await general graphics work, run on the main thread at the
start of each frame.

There are no longer separate ID types for static resources, these now
use the standard buffer and image IDs.

The main advantage to this approach is to handle resources that have
dependencies between each other, such as acceleration structures that
need buffer devices addresses. This simplifies areas of code that had
to manually split loading into multiple phases and call these phases
from the main thread. Now the loading code can await dependencies
itself, the main thread can simply wait for the fully created resource
to arrive.

The main disadvantage vs the previous system is that care must be taken
to add concurrency where necessary to keep loading fast. This means it
is better to spawn separate sub-tasks for independent resources and
await them as late as possible, rather than spawning and awaiting for
tasks immediately.